--- a/bin/plugin/commands/common.js
+++ b/bin/plugin/commands/common.js
@@ -1,26 +1,72 @@
 /**
  * External dependencies
  */
+const fs = require( 'fs' );
+const rimraf = require( 'rimraf' );
 const semver = require( 'semver' );
-const SimpleGit = require( 'simple-git' );
 
 /**
  * Internal dependencies
  */
-const { readJSONFile } = require( '../lib/utils' );
+const { log, formats } = require( '../lib/logger' );
+const { runStep, readJSONFile } = require( '../lib/utils' );
+const git = require( '../lib/git' );
+const config = require( '../config' );
+
+/**
+ * Clone the repository and returns the working directory.
+ *
+ * @param {string} abortMessage Abort message.
+ *
+ * @return {Promise<string>} Repository local path.
+ */
+async function runGitRepositoryCloneStep( abortMessage ) {
+	// Cloning the repository.
+	let gitWorkingDirectoryPath;
+	await runStep( 'Cloning the Git repository', abortMessage, async () => {
+		log( '>> Cloning the Git repository' );
+		gitWorkingDirectoryPath = await git.clone( config.gitRepositoryURL );
+		log(
+			'>> The Git repository has been successfully cloned in the following temporary folder: ' +
+				formats.success( gitWorkingDirectoryPath )
+		);
+	} );
+
+	return gitWorkingDirectoryPath;
+}
+
+/**
+ * Clean the working directories.
+ *
+ * @param {string[]} folders      Folders to clean.
+ * @param {string}   abortMessage Abort message.
+ */
+async function runCleanLocalFoldersStep( folders, abortMessage ) {
+	await runStep( 'Cleaning the temporary folders', abortMessage, async () => {
+		await Promise.all(
+			folders.map( async ( directoryPath ) => {
+				if ( fs.existsSync( directoryPath ) ) {
+					await rimraf( directoryPath, ( err ) => {
+						if ( err ) {
+							throw err;
+						}
+					} );
+				}
+			} )
+		);
+	} );
+}
 
 /**
  * Finds the name of the current plugin release branch based on the version in
- * the package.json file and the latest `trunk` branch in `git`.
+ * the package.json file.
  *
  * @param {string} gitWorkingDirectoryPath Path to the project's working directory.
  *
  * @return {string} Name of the plugin release branch.
  */
 async function findPluginReleaseBranchName( gitWorkingDirectoryPath ) {
-	await SimpleGit( gitWorkingDirectoryPath )
-		.fetch( 'origin', 'trunk' )
-		.checkout( 'trunk' );
+	await git.checkoutRemoteBranch( gitWorkingDirectoryPath, 'trunk' );
 
 	const packageJsonPath = gitWorkingDirectoryPath + '/package.json';
 	const mainPackageJson = readJSONFile( packageJsonPath );
@@ -95,4 +141,6 @@ function calculateVersionBumpFromChangelog(
 module.exports = {
 	calculateVersionBumpFromChangelog,
 	findPluginReleaseBranchName,
+	runGitRepositoryCloneStep,
+	runCleanLocalFoldersStep,
 };