--- a/bin/plugin/commands/performance.js
+++ b/bin/plugin/commands/performance.js
@@ -241,7 +241,7 @@ async function runPerformanceTests( branches, options ) {
 
 	log( '    >> Installing dependencies and building packages' );
 	await runShellScript(
-		'npm ci && node ./bin/packages/build.js',
+		'npm ci && npm run build:packages',
 		performanceTestDirectory
 	);
 	log( '    >> Creating the environment folders' );
@@ -275,10 +275,7 @@ async function runPerformanceTests( branches, options ) {
 			await SimpleGit( buildPath ).reset( 'hard' ).checkout( branch );
 
 			log( `        >> Building the ${ fancyBranch } branch` );
-			await runShellScript(
-				'npm ci && node ./bin/packages/build.js',
-				buildPath
-			);
+			await runShellScript( 'npm ci && npm run build', buildPath );
 		}
 
 		await runShellScript(