--- a/bin/tsconfig.json
+++ b/bin/tsconfig.json
@@ -25,6 +25,7 @@
 		"./plugin/lib/version.js",
 		"./plugin/lib/logger.js",
 		"./plugin/lib/utils.js",
+		"./plugin/lib/git.js",
 		"./validate-package-lock.js"
 	]
 }