> **Note**: **DO NOT MERGE** this is a long-running branch that needs to remain an open PR (and not a draft) so that it runs the normal CI test suite. It's purpose is for data-gathering to measure the impact of a sequence of changes to the performance test suite.

## What

This branch is intended to follow `trunk` without the set of changes meant improve the runtime performance of the Performance Test CI workflow.

![branch-compare-44907-45175](https://user-images.githubusercontent.com/5431237/203651787-1cb56bce-5837-4744-82a5-ea44194a21d8.png)

 - average time to run performance test workflow in `trunk` as of this test is **20.6min**
 - average time to run performance test workflow without the changes is **53.3min**
 - measured reduction in average runtime is therefore **-32.7min ✅**
 - measured against 181 test runs in `trunk` and 76 test runs in this branch (there are more in `trunk` because the tests finish faster and I'm able to get more data in the same amount of time)
 - standard deviation in `trunk` is **2.6min** vs. without the changes at **7.5min**, so not only has the average test time reduced but also the extrema for the test runs has also reduced and the runtimes are now more consistent.
 - there is overwhelming statistical reason to believe that the reduction is real. the chance that random sampling variation could account for these changes is almost zero (P-value is on the order of 2·10<sup>-16</sup>)

### Results

| Variant | N | Mean / Median | Minimum | StdDev | Δ w/o changes | Δ trunk |
|---|---|---|---|---|---|---|
| Without changes #45175 | 30 | 50m 43s / 49m 3s | 42m 22s | 6m 56s | 0m 0s 100% | 5m 33s 112% |
| Empty change #44907 | 30 | 45m 10s / 43m 24s | 32m 19s | 7m 33s | -5m 33s 89.0% | 0m 0s 100% |
| 🚫 --runInBand #44905 | 30 | 48m 33s / 43m 52s | 38m 16s | 10m 14s | -2m 10s 95.7% | 3m 22s 107% |
| ✅ --depth=2 at merge branch #45057 | 30 | 42m 47s / 42m 27s | 31m 7s | 5m 3s | -7m 56s 84.3% | -2m 22s 94.7% |
| ✅ No-generate-types #45284 | 30 | 45m 54s / 45m 4s | 37m 49s | 5m 40s | -4m 49s 90.5% | 0m 44s 102% |
| 🚫 Clone local checkout #45188 | 29 | 46m 50s / 45m 6s | 38m 7s | 6m 29s | -3m 53s 92.3% | 1m 40s 104% |
| Refactor use of ENV #45255 | 30 | 49m 56s / 50m 2s | 38m 47s | 7m 27s | -0m 47s 98.4% | 4m 45s 111% |
| ✅ Reuse tests-branch build #45737 | | | | | |
| 🚫 Single fetch with `refspecs` #45780 | | | | | |

In order to eliminate some bias introduced by variations in test run durations over time these numbers reflect only the 30 most recent test runs.

**Nov. 1**

I'm rebasing all branches and am going to let them re-gather data to see if any clearer patterns emerge. Will be ignoring all test run data before Nov. 1 to see if there were effects due to getting too far behind `trunk`.


## Why

By comparing the runtimes in this branch against those in `trunk` we can measure the impact of the overall optimization project while deploying improvements incrementally. Running this over time will not only eliminate sampling bias and noise, but it will more importantly help us to eliminate the impact of other changes on the performance test suite which might be conflated with changes we're making to the suite itself (e.g. if someone adds a new set of performance tests to the suite which would slow them down).