@dmsnell 2023-02-10
Closing for now since active work on the performance tests performance is on hold.